Local copy of JQuery and JQueryUI.
The JQueryUI dump isn't customised at all at the moment. Can trim it
down once we know which bits we are and aren't using.

Using a local version, rather than the Google CDN version for local
development purposes (Google requires an API key which is tied to a
particular external URL).