Create model of all supported languages.
This helps in a number of places where we want to allow the user (or
code) to select one from a valid list.

The change propogates through a lot of because Word.language is now
Word.lang_code.code and there's a fairly long migration chain in place
that works forwards and backwards to keep the data consistent.