Change UserProfile to Profile class.
Closes #18.

We forgot to make the link from UserProfile -> User be the attribute
"user". This fixes that, but the easiest way to migrate between models
when changing the primary key is to create a new model and copy the data
back and forth. So I've created the Profile model that mirrors the old
UserProfile, except that "student" is now called "user".